memecah kode main.dart menjadi beberapa komponen agar kode tidak terlalu panjang
